Move search apps from __init__.py to separate module
