+++ datahub/search/contact/__init__.py
@@ -1,16 +1,3 @@
-from datahub.company.models import Contact as DBContact
+from .apps import ContactSearchApp
 
-from .models import Contact
-from .views import SearchContactAPIView
-
-from ..apps import SearchApp
-
-
-class ContactSearchApp(SearchApp):
-    """SearchApp for contacts"""
-
-    name = 'contact'
-    plural_name = 'contacts'
-    ESModel = Contact
-    DBModel = DBContact
-    view = SearchContactAPIView
+__all__ = ('ContactSearchApp',)
