+++ datahub/search/contact/apps.py
@@ -1,0 +1,16 @@
+from datahub.company.models import Contact as DBContact
+
+from .models import Contact
+from .views import SearchContactAPIView
+
+from ..apps import SearchApp
+
+
+class ContactSearchApp(SearchApp):
+    """SearchApp for contacts"""
+
+    name = 'contact'
+    plural_name = 'contacts'
+    ESModel = Contact
+    DBModel = DBContact
+    view = SearchContactAPIView
